Add files via upload
An interface is a class-like construct that contains only constants and abstract methods. 
In many ways an interface is similar to an abstract class, but its intent is to specify common 
behavior for objects of related classes or unrelated classes. For example, using appropriate 
interfaces, you can specify that the objects are comparable, edible, and/or cloneable. 
To distinguish an interface from a class, Java uses the following syntax to define an interface:
modifier interface InterfaceName {
/** Constant declarations */
/** Abstract method signatures */
}